ci: fix all Build CI jobs (fmt, libos, bare-metal aarch64/riscv64, workspace)
The Build CI workflow had been red on master for many consecutive runs.
This restores it across every job, verified locally against each job's
exact commands.

workspace (fmt + build + clippy --all-features):
- cargo fmt --all over the legacy files that predated the pinned
  nightly-2026-05-01 rustfmt.
- Resolve every clippy lint the pinned toolchain now raises under the
  crates' deny(warnings): mechanical fixes (is_multiple_of, div_ceil,
  strip_prefix/suffix, sort_by_key, matches!, contains_key, size_of_val,
  inline const array init, legacy numeric methods, unnecessary casts,
  collapsible if/if-let, redundant closures, …) plus a small number of
  targeted #[allow]s with rationale for intentional low-level patterns
  (register-offset tables, MMIO transmutes/packed structs, per-CPU
  UnsafeCell aliasing, wide hardware bring-up signatures, a genuinely
  single-region Vec).

test-libos (linux, zircon):
- Add a hosted (non-bare) no-op `report_stuck`/`set_deadlock_hook` to
  vendor/kernel-sync so vendor/PreemptiveScheduler compiles off bare
  metal — the error that broke the whole libos build.

test-bare-metal (aarch64, riscv64):
- Gate the x86_64-only NVIDIA RM stack (nvidia-rm-sys shim + the drivers
  display::nvidia module and its PCI registration) to target_arch =
  "x86_64", with a cross-arch `boot_edid()` fallback. It was compiling
  an x86 C object and x86 register references into the ARM/RISC-V link.
- Gate x86_64-only syscalls (inotify_init, signalfd, getpgrp,
  eclipse_dns_query) and the x86 crash-register dump to x86_64 in the
  dispatcher / loader.
- Fix pre-existing riscv breakage: wrong Arc import and a super:: path
  to kernel_page_table in kernel-hal.
- Fix per-arch dead_code / unused-import / lint errors surfaced once the
  earlier failures were cleared.

Makefile: skip the initramfs copy when no rootfs image is present, so a
build-only run (CI compiles the kernel without a rootfs) does not die on
`cp x86_64.img`.

Verified locally: workspace fmt/build/clippy --all-features; both libos
modes build+clippy; aarch64 and riscv64 bare build+clippy+xtask; the
x86_64 zircon build compiles fully (its final link needs the NVIDIA RM
submodule objects present only in CI, past which the initramfs fix
applies). No functional/behavioral changes to the kernel.

+++ linux-syscall/src/lib.rs
@@ -232,6 +232,9 @@ impl Syscall<'_> {
                     .await
             }
             Sys::EVENTFD2 => self.sys_eventfd2(a0 as u32, a1),
+            // Legacy `inotify_init` exists only in the x86_64 table; the generic
+            // ABI (aarch64/riscv64) provides only `inotify_init1`.
+            #[cfg(target_arch = "x86_64")]
             Sys::INOTIFY_INIT => self.sys_inotify_init1(0),
             Sys::INOTIFY_INIT1 => self.sys_inotify_init1(a0),
             Sys::INOTIFY_ADD_WATCH => self.sys_inotify_add_watch(a0, a1.into(), a2 as u32),
@@ -241,6 +244,8 @@ impl Syscall<'_> {
             Sys::TIMERFD_SETTIME => self.sys_timerfd_settime(a0.into(), a1, a2.into(), a3.into()),
             Sys::TIMERFD_GETTIME => self.sys_timerfd_gettime(a0.into(), a1.into()),
             Sys::SIGNALFD4 => self.sys_signalfd4(a0.into(), a1.into(), a2, a3),
+            // Legacy `signalfd` is x86_64-only; the generic ABI has `signalfd4`.
+            #[cfg(target_arch = "x86_64")]
             Sys::SIGNALFD => self.sys_signalfd4(a0.into(), a1.into(), a2, 0),
 
             Sys::SOCKETPAIR => self.sys_socketpair(a0, a1, a2, a3.into()),
@@ -432,6 +437,8 @@ impl Syscall<'_> {
             // an interactive busybox `sh` cannot determine its own process group
             // during job-control setup, takes the "I am a background job" branch
             // and `kill(0, SIGTTIN)`s itself — which then terminated the shell.
+            // Legacy `getpgrp` is x86_64-only; the generic ABI uses getpgid(0).
+            #[cfg(target_arch = "x86_64")]
             Sys::GETPGRP => self.sys_getpgid(0),
             Sys::GETGROUPS => self.sys_getgroups(a0, a1.into()),
             Sys::SETGROUPS => self.sys_setgroups(a0, a1.into()),
@@ -462,7 +469,7 @@ impl Syscall<'_> {
             Sys::MEMBARRIER => self.sys_membarrier(a0 as i32, a1 as u32, a2 as i32),
             Sys::PRLIMIT64 => self.sys_prlimit64(a0, a1, a2.into(), a3.into()),
             Sys::REBOOT => self.sys_reboot(a0 as u32, a1 as u32, a2 as u32, a3.into()),
-            Sys::GETRANDOM => self.sys_getrandom(a0.into(), a1 as usize, a2 as u32),
+            Sys::GETRANDOM => self.sys_getrandom(a0.into(), a1, a2 as u32),
             Sys::STATX => self.sys_statx(a0.into(), a1.into(), a2, a3 as u32, a4.into()),
             Sys::RT_SIGQUEUEINFO => self.unimplemented("rt_sigqueueinfo", Ok(0)),
 
@@ -484,6 +491,8 @@ impl Syscall<'_> {
             //            Sys::DELETE_MODULE => self.sys_delete_module(a0.into(), a1 as u32),
             #[cfg(not(target_arch = "aarch64"))]
             Sys::BLOCK_IN_KERNEL => self.sys_block_in_kernel(),
+            // Custom `eclipse_dns_query` is only in the x86_64 syscall table.
+            #[cfg(target_arch = "x86_64")]
             Sys::ECLIPSE_DNS_QUERY => self.sys_eclipse_dns_query(a0.into(), a1, a2, a3.into(), a4),
             Sys::PERF_EVENT_OPEN => {
                 self.sys_perf_event_open(a0, a1 as i32, a2 as i32, a3 as i32, a4)
